+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/domain/handlers/common/DmnStringValue.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.wacaseeventhandler.domain.handler.common;
+package uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.common;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
