+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/services/CaseEventHandler.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.wacaseeventhandler.services;
 
-import uk.gov.hmcts.reform.wacaseeventhandler.domain.handler.common.EvaluateResponse;
-import uk.gov.hmcts.reform.wacaseeventhandler.domain.handler.common.EventInformation;
+import uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.common.EvaluateResponse;
+import uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.common.EventInformation;
 
 import java.util.List;
