+++ src/test/java/uk/gov/hmcts/reform/wacaseeventhandler/services/WarningTaskHandlerTest.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.wacaseeventhandler.services;
 
 import org.junit.jupiter.api.Test;
-import uk.gov.hmcts.reform.wacaseeventhandler.domain.handler.common.EventInformation;
+import uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.common.EventInformation;
 
 import static org.assertj.core.api.Assertions.assertThat;
